// FrontendForTransfer/src/models/Cart/ICartItem.ts
export default interface ICartItem {
    quantity: number;
    id: number;
    code: string;

    fromCityName: string;
    fromCountryName: string;

    toCityName: string;
    toCountryName: string;

    departureTime: string;
    arrivalTime: string;

    seatsTotal: number;
    seatsAvailable: number;

    statusName: string;
}